% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_reads.R
\name{cap_reads}
\alias{cap_reads}
\title{cap_reads}
\usage{
cap_reads(bam_file, max_reads, ncores = 1, out_dir = dirname(bam_file))
}
\arguments{
\item{bam_file}{an indexed bam file}

\item{max_reads}{target number of reads}

\item{ncores}{number of cores to use for samtools view operation}
}
\value{
path to subsampled file
}
\description{
create a subsampled version of bam_file that is approximately the number specified by max_reads
}
\examples{
cap_reads
}
