% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{load_overlap_convergence}
\alias{load_overlap_convergence}
\alias{plot_overlap_convergence_bars}
\alias{plot_overlap_convergence_trends}
\title{load_overlap_convergence}
\usage{
load_overlap_convergence(psc, a, b, min_signalValue = 1)

plot_overlap_convergence_bars(olap_dt, a, b)

plot_overlap_convergence_trends(olap_dt, a, b)
}
\arguments{
\item{psc}{A valid peaksat_config object. Create using the \link[peaksat]{peaksat_config} function.}

\item{min_signalValue}{}

\item{olap_dt}{output from load_overlap_convergence function}
}
\value{
ggplot showing stacked bars of overlaps at different subsample levels.

ggplot showing line plots of overlap trends at different subsample levels.
}
\description{
load_overlap_convergence

plot_overlap_convergence_bars

plot_overlap_convergence_trends
}
\examples{
psc = peaksat_config.example()
cnv_dt = load_convergence(psc, "rep1", "rep2")
plot_convergence_bars(cnv_dt, "rep1", "rep2")
plot_convergence_trends(cnv_dt, "rep1", "rep2")
}
