% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_counts.R
\name{.general_load_info.wd}
\alias{.general_load_info.wd}
\title{.general_load_info.wd}
\usage{
.general_load_info.wd(wds, min_signalValue = 1, type = c("count", "width")[1])
}
\arguments{
\item{wds}{directories to load count data from.}
}
\value{
data.table of peak count vs read count.
}
\description{
.general_load_info.wd
}
