% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_depth.R
\name{estimate_depth.linear}
\alias{estimate_depth.linear}
\title{Title}
\usage{
estimate_depth.linear(
  cnt_dt,
  target_peaks = NULL,
  min_read_count = 5e+06,
  max_read_count = 1e+08
)
}
\arguments{
\item{max_read_count}{}
}
\value{

}
\description{
Title
}
\examples{
psc = peaksat_config("~/R/peaksat_paper/peaksat_outputs/peak_saturation_10A_H4ac_seq1_only")
cnt_dt = load_counts(psc)
est_res = estimate_depth.linear(cnt_dt[grepl("H4K8AC", sample)], target_peaks = 26e3, min_read_count = 10e6)
est_res$plots
est_res$estimates
est_res$estimates[, .(sample, target_Mreads = saturation_read_count/1e6, current_Mreads = current_read_count/1e6, needed_Mreads = saturation_read_count/1e6-current_read_count/1e6)]
}
