% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_depth.R
\name{estimate_depth.log}
\alias{estimate_depth.log}
\title{Estimate read depth required to reach target peaks based on logistic regression.}
\usage{
estimate_depth.log(
  cnt_dt,
  target_peaks = NULL,
  min_read_count = 5e+06,
  max_read_count = 1e+08
)
}
\arguments{
\item{cnt_dt}{data.table from load_counts}

\item{target_peaks}{The estimated number of peaks at saturation.}

\item{min_read_count}{Read counts below this value are not used when fitting regression}

\item{max_read_count}{Read counts above this value are not used when fitting regression}
}
\value{
A named list. Item 1 is a data.table containing estimates for saturation. Item 2 is a list of plots showing regression curves.
}
\description{
Estimate read depth required to reach target peaks based on logistic regression.
}
