% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_counts.R
\name{load_counts.wd}
\alias{load_counts.wd}
\title{load_counts.wd}
\usage{
load_counts.wd(wds)
}
\arguments{
\item{wds}{directories to load count data from.}
}
\value{
data.table of peak count vs read count.
}
\description{
load_counts.wd
}
