% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_counts.R
\name{load_peaks}
\alias{load_peaks}
\title{load_peaks}
\usage{
load_peaks(psc, min_signalValue = 1, selection = NULL)
}
\arguments{
\item{psc}{peaksat_config object}

\item{min_signalValue}{}
}
\value{
list of GRanges of peaks filtered by min_signalValue.
}
\description{
load_peaks
}
