% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_counts.R
\name{load_widths}
\alias{load_widths}
\title{load_widths}
\usage{
load_widths(psc, min_signalValue = 1)
}
\arguments{
\item{psc}{peaksat_config object}

\item{min_signalValue}{}
}
\value{
data.table of peak count vs read count.
}
\description{
load_widths
}
