% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_pool.R
\name{make_meta_pool_cmd}
\alias{make_meta_pool_cmd}
\title{make_pool_cmd}
\usage{
make_meta_pool_cmd(psc, bams, name, pool_script = get_pool_script())
}
\arguments{
\item{pool_script}{}
}
\value{

}
\description{
make_pool_cmd
}
