% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_peak_saturation.R
\name{make_submit_cmd}
\alias{make_submit_cmd}
\title{Title}
\usage{
make_submit_cmd(pc, treatment, control, submit_script = get_submit_script())
}
\value{

}
\description{
Title
}
\examples{
treat_bam = "/slipstream/galaxy/uploads/working/qc_framework/output/MCF10A_H3K4AC_R1/MCF10A_H3K4AC_R1.bam"
ctrl_bam = "/slipstream/galaxy/uploads/working/qc_framework/output/MCF10A_input_R1/MCF10A_input_R1.bam"
pc = peaksat_config()
cmd = make_submit_cmd(pc, treat_bam, ctrl_bam)
sub_out = system(cmd, intern = TRUE)
sub_out
}
