% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaksat_config.R
\docType{class}
\name{peaksat_config-class}
\alias{peaksat_config-class}
\title{Title}
\value{

}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{stat}}{character.}

\item{\code{stat_value}}{numeric.}

\item{\code{is_PE}}{logical.}

\item{\code{out_dir}}{character.}

\item{\code{macs2_path}}{character.}

\item{\code{samtools_path}}{character.}

\item{\code{submit_script}}{character.}

\item{\code{job_scheduler}}{character.}

\item{\code{noModel}}{logical}
}}

