% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaksat_config.R
\name{peaksat_config.example}
\alias{peaksat_config.example}
\title{peaksat_config.example}
\usage{
peaksat_config.example()
}
\value{
A peaksat_config object using data packaged with peaksat.
}
\description{
peaksat_config.example
}
\examples{
psc = peaksat_config.example()
}
