% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recount.R
\name{recount_peak}
\alias{recount_peak}
\title{recount_peak}
\usage{
recount_peak(psc, min_signalValues = c(1, 2, 5, 10, 20, 50))
}
\arguments{
\item{min_signalValues}{}
}
\value{

}
\description{
recount_peak
}
