% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_peaksat_jobs.R
\name{submit_peaksat_jobs}
\alias{submit_peaksat_jobs}
\title{Title}
\usage{
submit_peaksat_jobs(
  psc,
  treat_bams,
  ctrl_bams,
  hold_jid_map = NULL,
  await_completion = TRUE
)
}
\arguments{
\item{psc}{A valid peaksat_config object}

\item{treat_bams}{Vector of ChIPseq bam files}

\item{ctrl_bams}{Vector of input bam files parallel to ChIPseq. If NULL, macs2 will be called without using an input file.}

\item{await_completion}{}
}
\description{
Title
}
