% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_pool_jobs.R
\name{submit_pool_jobs}
\alias{submit_pool_jobs}
\title{Title}
\usage{
submit_pool_jobs(
  psc,
  bam_groups,
  bam_group_names = names(bam_groups),
  pool_script = get_pool_script()
)
}
\arguments{
\item{pool_script}{}
}
\description{
Title
}
