% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_depth.R
\name{estimate_depth.linear}
\alias{estimate_depth.linear}
\title{Title}
\usage{
estimate_depth.linear(
  cnt_dt,
  target_peaks = NULL,
  min_read_count = 5e+06,
  max_read_count = 1e+08
)
}
\arguments{
\item{max_read_count}{}
}
\value{

}
\description{
Title
}
