% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_counts.R
\name{load_counts}
\alias{load_counts}
\title{load_counts}
\usage{
load_counts(psc, min_signalValue = 1, selection = NULL)
}
\arguments{
\item{psc}{peaksat_config object}

\item{min_signalValue}{}
}
\value{
data.table of peak count vs read count.
}
\description{
load_counts
}
