% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaksat_config.R
\name{peaksat_config}
\alias{peaksat_config}
\title{peaksat_config}
\usage{
peaksat_config(
  out_dir = getOption("PS_OUTDIR", file.path(getwd(), "peak_saturation")),
  stat = valid_stats$qValue,
  stat_value = 0.01,
  is_PE = FALSE,
  macs2_path = get_macs2_path(),
  samtools_path = get_samtools_path(),
  submit_script = get_submit_script(),
  job_scheduler = "SGE",
  noModel = FALSE
)
}
\arguments{
\item{out_dir}{}
}
\description{
peaksat_config
}
\examples{
psc = peaksat_config()
call_submit_script(psc, "treat.bam", "ctrl.bam")
}
