% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_peak_saturation.R
\name{peaksat_config}
\alias{peaksat_config}
\title{Title}
\usage{
peaksat_config(
  stat = valid_stats$qValue,
  stat_value = 0.01,
  is_PE = FALSE,
  out_dir = getOption("PS_OUTDIR", file.path(getwd(), "peak_saturation")),
  macs2_path = get_macs2_path(),
  submit_script = get_submit_script()
)
}
\arguments{
\item{out_dir}{}
}
\value{

}
\description{
Title
}
\examples{
pc = peaksat_config()
call_submit_script(pc, "treat.bam", "ctrl.bam")
}
