% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_pool.R
\name{submit_meta_pool_jobs}
\alias{submit_meta_pool_jobs}
\title{Title}
\usage{
submit_meta_pool_jobs(
  pc,
  bam_groups,
  bam_group_names = names(bams),
  pool_script = get_pool_script(),
  skip_peaksat = FALSE
)
}
\arguments{
\item{skip_peaksat}{}
}
\value{

}
\description{
Title
}
