% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_peaksat_jobs.R
\name{submit_peaksat_jobs}
\alias{submit_peaksat_jobs}
\title{Title}
\usage{
submit_peaksat_jobs(
  psc,
  treat_bams,
  ctrl_bams,
  hold_jid_map = NULL,
  await_completion = TRUE
)
}
\arguments{
\item{await_completion}{}
}
\value{

}
\description{
Title
}
