% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_peak_saturation.R
\name{submit_peaksat_jobs}
\alias{submit_peaksat_jobs}
\title{Title}
\usage{
submit_peaksat_jobs(pc, treat_bams, ctrl_bams, await_completion = TRUE)
}
\arguments{
\item{await_completion}{}
}
\value{

}
\description{
Title
}
